requirements.txt
aiosignal==1.3.2
attrs==25.3.0
boto3==1.37.9
botocore==1.37.9
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
decorator==5.1.1
filelock==3.17.0
frozenlist==1.6.0
fsspec==2025.3.0
huggingface-hub==0.29.2
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.3
opensearch-py==2.5.0
packaging==24.0
pluggy==1.5.0
protobuf==6.30.2
py==1.11.0
Pygments==2.19.1
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
ray==2.44.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.1
requests-aws4auth==1.3.1
retry==0.9.2
rich==14.0.0
rpds-py==0.24.0
s3transfer==0.11.4
safetensors==0.5.3
scipy==1.15.2
six==1.16.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==1.26.18
wonderwords==2.2.0
locust==2.32.4
gevent==24.11.1
